Resize S1 image placeholder & text container
The image's width of 400px was more than the container's width; thus, the GIF overlapped with the text container on its left.

Currently, it does not seem like a problem due to lack of text next to it. However, if there were more text, the GIF would have covered some of it.